Refactor. separate components
